requirements.txt
python-binance
typer[all]
rich
python-dotenv
requests
